Rework of thread locking/unlocking for NativeAV
Former-commit-id: 3f2ddff3613a7befccebdf530eeeeb2a5a838d06 [formerly 24daf9290dfb343eea30825ceb46db5aaba7d893]
Former-commit-id: 791fd9def15aac559a47c3655d5fba7e297f8a28

+++ av/NativeAudioNode.h
@@ -207,7 +207,7 @@ class NativeAudioNodeDelay : public NativeAudioNode
             DELAY, WET, DRY 
         };
 
-        double delay;
+        int delay;
         double wet;
         double dry;
         float **buffers;
